support dynamic configuration
restic-helper now acts as a wrapper for the other scripts.

If an executable named "before" exists in the profile directory; it
will be run before the restic action. Any output from the script will
be captured and evaluated in a `set -o allexport` context.

If an executable named "after" exists, it will be run after the restic
action if the action completes successfully.